+++ Modules/SWMB/LocalMachine-Privacy.psm1
@@ -1101,6 +1101,32 @@ Function TweakViewBackupMessages { # RESINFO
 }
 
 ################################################################
+
+# This policy setting lets you turn off cloud optimized content in all Windows experiences.
+# https://admx.help/?Category=Windows_10_2016&Policy=Microsoft.Policies.CloudContent::DisableCloudOptimizedContent
+# Disable
+Function TweakDisableCloudOptimizedContent { # RESINFO
+	Write-Output "Turn off cloud optimized content..."
+	If (!(Test-Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows\CloudContent")) {
+		New-Item -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows\CloudContent" | Out-Null
+	}
+	Set-ItemProperty -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows\CloudContent" -Name "DisableCloudOptimizedContent" -Type DWord -Value 0
+}
+
+# Enable
+Function TweakEnableCloudOptimizedContent { # RESINFO
+	Write-Output "Turn on cloud optimized content..."
+	Remove-ItemProperty -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows\CloudContent" -Name "DisableCloudOptimizedContent" -ErrorAction SilentlyContinue
+}
+
+# View
+Function TweakViewCloudOptimizedContent { # RESINFO
+	Write-Output "Viewing cloud optimized content (not exist - enable, 0 disable)..."
+	Get-ItemProperty -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows\CloudContent" -Name "DisableCloudOptimizedContent" -ErrorAction SilentlyContinue
+}
+
+
+################################################################
 ###### Export Functions
 ################################################################