+++ NEWS.md
@@ -5,6 +5,8 @@
 The `wisemoui.ps1` graphical interface has been updated.
 There's now an L (Log) and E (Edit) button next to each scheduled task to view the last log file and edit the task's preset file.
 
+New presets/rules:
+* 2024/11/28 - `DisableCloudOptimizedContent`/`EnableCloudOptimizedContent`/`ViewCloudOptimizedContent` -> Turn off cloud optimized content in all Windows experiences
 
 ## Version 3.17 (2024/10/01)